Move Service and Context to their own package.
- `service` package holds service related struct and methods. It's just
  a move, stuff like `Namer` should be moved away from this package and
  refactor in a way they don't call the docker API (e.g).
- `ctx` holds the `Context` struct used to pass factory and stuff :)
