+++ cli/docker/app/commands.go
@@ -3,8 +3,8 @@ package app
 import (
 	"github.com/Sirupsen/logrus"
 	"github.com/docker/libcompose/cli/command"
-	"github.com/docker/libcompose/docker"
 	"github.com/docker/libcompose/docker/client"
+	"github.com/docker/libcompose/docker/ctx"
 	"github.com/urfave/cli"
 )
 
@@ -41,7 +41,7 @@ func DockerClientFlags() []cli.Flag {
 }
 
 // Populate updates the specified docker context based on command line arguments and subcommands.
-func Populate(context *docker.Context, c *cli.Context) {
+func Populate(context *ctx.Context, c *cli.Context) {
 	command.Populate(&context.Context, c)
 
 	context.ConfigDir = c.String("configdir")
