+++ cli/docker/app/factory.go
@@ -3,6 +3,7 @@ package app
 import (
 	"github.com/docker/libcompose/cli/logger"
 	"github.com/docker/libcompose/docker"
+	"github.com/docker/libcompose/docker/ctx"
 	"github.com/docker/libcompose/project"
 	"github.com/urfave/cli"
 )
@@ -13,7 +14,7 @@ type ProjectFactory struct {
 
 // Create implements ProjectFactory.Create using docker client.
 func (p *ProjectFactory) Create(c *cli.Context) (project.APIProject, error) {
-	context := &docker.Context{}
+	context := &ctx.Context{}
 	context.LoggerFactory = logger.NewColorLoggerFactory()
 	Populate(context, c)
 	return docker.NewProject(context, nil)
