+++ docker/service/name.go
@@ -1,4 +1,4 @@
-package docker
+package service
 
 import (
 	"fmt"
