+++ integration/api_test.go
@@ -6,6 +6,7 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/docker/libcompose/docker"
+	"github.com/docker/libcompose/docker/ctx"
 	"github.com/docker/libcompose/project"
 	"github.com/docker/libcompose/project/options"
 )
@@ -24,7 +25,7 @@ service:
   volumes:
     - /etc/selinux:/etc/selinux`
 
-	project, err := docker.NewProject(&docker.Context{
+	project, err := docker.NewProject(&ctx.Context{
 		Context: project.Context{
 			ComposeBytes: [][]byte{[]byte(service)},
 			ProjectName:  "test-volume-without-compose-file",
